+++ apps/filecoin-site/src/app/[locale]/case-studies/components/PageHeader.tsx
@@ -29,13 +29,13 @@ export function PageHeader({
         <SharedPageHeader {...pageHeaderProps} />
       </div>
 
-      {image && (
-        <div className="relative aspect-video w-full object-cover">
+      {image && image.src && (
+        <div className="relative aspect-video">
           <Image
             fill
             src={image.src}
             alt={image.alt || ''}
-            className="rounded-xl"
+            className="rounded-2xl object-cover"
           />
         </div>
       )}