+++ packages/ui-filecoin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @filecoin-foundation/ui-filecoin
 
+## 0.5.10
+
+### Patch Changes
+
+- 590001d: Add new capitalize util function and refactor export
+- f963e9f: Refactor CardImage and types
+
+## 0.5.9
+
+### Patch Changes
+
+- ffc4ad8: Refactor and review PageSection spacings
+- 65979ec: Add highContrast variant to PageHeader
+
 ## 0.5.8
 
 ### Patch Changes